fix: Ensure better sync between quest state and QuestStepPanel
This makes use of the actual active sidebar step for the queststeppanel, rather than a local variable representing it based on a last-known state.

This should ensure that panels are properly checked, and there's no chance of this variable not representing the actual quest state.

This does mean recolour of the sidebar can occur on things which are already set, but I don't think this should have any noted performance changes.